Revert "Rename extenstion to ddflareextension (#28390)"
This reverts commit 14a8524c8d843adb10d4e85be1604e33c084d259.
