+++ comp/README.md
@@ -383,9 +383,9 @@ Package configstore defines the otel agent configstore component.
 
 Package converter defines the otel agent converter component.
 
-### [comp/otelcol/ddflareextension](https://pkg.go.dev/github.com/DataDog/datadog-agent/comp/otelcol/ddflareextension)
+### [comp/otelcol/extension](https://pkg.go.dev/github.com/DataDog/datadog-agent/comp/otelcol/extension)
 
-Package ddflareextension defines the OpenTelemetry Extension component.
+Package extension defines the OpenTelemetry Extension component.
 
 ### [comp/otelcol/logsagentpipeline](https://pkg.go.dev/github.com/DataDog/datadog-agent/comp/otelcol/logsagentpipeline)
