+++ comp/otelcol/converter/impl/testdata/extensions/other-extensions/config-result.yaml
@@ -1,8 +1,8 @@
 receivers:
-  otlp:
+    otlp:
 
 exporters:
-  nop:
+  nop: 
 
 extensions:
   otlp_encoding/user-defined:
@@ -11,25 +11,18 @@ extensions:
   health_check/dd-autoconfigured:
   zpages/dd-autoconfigured:
     endpoint: "localhost:55679"
-  ddflare/dd-autoconfigured:
+  datadog/dd-autoconfigured:
+    
 
 service:
-  extensions:
-    [
-      otlp_encoding/user-defined,
-      pprof/dd-autoconfigured,
-      zpages/dd-autoconfigured,
-      health_check/dd-autoconfigured,
-      ddflare/dd-autoconfigured,
-    ]
-  pipelines:
-    traces:
-      receivers: [nop]
-      exporters: [nop]
-    metrics:
-      receivers: [nop]
-      exporters: [nop]
-    logs:
-      receivers: [nop]
-      exporters: [nop]
-
+    extensions: [otlp_encoding/user-defined, pprof/dd-autoconfigured, zpages/dd-autoconfigured, health_check/dd-autoconfigured, datadog/dd-autoconfigured]
+    pipelines:
+        traces:
+            receivers: [nop]
+            exporters: [nop]
+        metrics:
+            receivers: [nop]
+            exporters: [nop]
+        logs:
+            receivers: [nop]
+            exporters: [nop]
