+++ comp/otelcol/converter/impl/testdata/processors/no-processor-partial/config.yaml
@@ -22,10 +22,10 @@ extensions:
   health_check/user-defined:
   zpages/user-defined:
     endpoint: "localhost:55679"
-  ddflare/user-defined:
+  datadog/user-defined:
    
 service:
-    extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+    extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, datadog/user-defined]
     pipelines:
         traces:
             receivers: [nop]
