+++ comp/otelcol/converter/impl/testdata/processors/other-processors/config-result.yaml
@@ -18,14 +18,14 @@ extensions:
   health_check/user-defined:
   zpages/user-defined:
     endpoint: "localhost:55679"
-  ddflare/user-defined:
+  datadog/user-defined:
     
 processors:
   infraattributes/dd-autoconfigured:
   k8sattributes:
 
 service:
-    extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+    extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, datadog/user-defined]
     pipelines:
         traces:
             receivers: [nop]
