+++ comp/otelcol/converter/impl/testdata/receivers/job-name-change/config-result.yaml
@@ -1,46 +1,40 @@
 receivers:
-  otlp:
-  prometheus/user-defined:
-    config:
-      scrape_configs:
-        - job_name: "datadog-agent"
-          scrape_interval: 10s
-          static_configs:
-            - targets: ["0.0.0.0:8888"]
+    otlp:
+    prometheus/user-defined:
+      config:
+        scrape_configs:
+          - job_name: 'datadog-agent'
+            scrape_interval: 10s
+            static_configs:
+              - targets: ['0.0.0.0:8888']
 
 exporters:
-  datadog:
-    api:
-      key: 12345
+    datadog:
+      api:
+        key: 12345
 
 extensions:
   pprof/user-defined:
   health_check/user-defined:
   zpages/user-defined:
     endpoint: "localhost:55679"
-  ddflare/user-defined:
-
+  datadog/user-defined:
+    
 processors:
   infraattributes/user-defined:
 
 service:
-  extensions:
-    [
-      pprof/user-defined,
-      zpages/user-defined,
-      health_check/user-defined,
-      ddflare/user-defined,
-    ]
-  pipelines:
-    traces:
-      receivers: [nop]
-      processors: [infraattributes/user-defined]
-      exporters: [datadog]
-    metrics:
-      receivers: [nop, prometheus/user-defined]
-      processors: [infraattributes/user-defined]
-      exporters: [datadog]
-    logs:
-      receivers: [nop]
-      processors: [infraattributes/user-defined]
-      exporters: [datadog]
+    extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, datadog/user-defined]
+    pipelines:
+        traces:
+            receivers: [nop]
+            processors: [infraattributes/user-defined]
+            exporters: [datadog]
+        metrics:
+            receivers: [nop, prometheus/user-defined]
+            processors: [infraattributes/user-defined]
+            exporters: [datadog]
+        logs:
+            receivers: [nop]
+            processors: [infraattributes/user-defined]
+            exporters: [datadog]
