+++ comp/otelcol/converter/impl/testdata/receivers/multi-dd-partial-prom/config.yaml
@@ -24,7 +24,7 @@ extensions:
   health_check/user-defined:
   zpages/user-defined:
     endpoint: "localhost:55679"
-  ddflare/user-defined:
+  datadog/user-defined:
 
 processors:
   infraattributes/user-defined:
@@ -33,7 +33,7 @@ service:
   telemetry:
     metrics:
       address: "localhost:1234"
-  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, datadog/user-defined]
   pipelines:
       traces:
           receivers: [nop]
