+++ comp/otelcol/converter/impl/testdata/receivers/no-changes/config.yaml
@@ -18,13 +18,13 @@ extensions:
   health_check/user-defined:
   zpages/user-defined:
     endpoint: "localhost:55679"
-  ddflare/user-defined:
+  datadog/user-defined:
     
 processors:
   infraattributes/user-defined:
 
 service:
-    extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+    extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, datadog/user-defined]
     pipelines:
         traces:
             receivers: [nop]
