+++ comp/otelcol/converter/impl/testdata/receivers/no-prom-not-default-addr/config.yaml
@@ -11,7 +11,7 @@ extensions:
   health_check/user-defined:
   zpages/user-defined:
     endpoint: "localhost:55679"
-  ddflare/user-defined:
+  datadog/user-defined:
     
 processors:
   infraattributes/user-defined:
@@ -20,7 +20,7 @@ service:
   telemetry:
     metrics:
       address: "localhost:1234"
-  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, ddflare/user-defined]
+  extensions: [pprof/user-defined, zpages/user-defined, health_check/user-defined, datadog/user-defined]
   pipelines:
       traces:
           receivers: [nop]
