+++ comp/otelcol/extension/impl/extension.go
@@ -3,8 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024-present Datadog, Inc.
 
-// Package ddflareextensionimpl defines the OpenTelemetry Extension implementation.
-package ddflareextensionimpl
+// Package extensionimpl defines the OpenTelemetry Extension implementation.
+package extensionimpl
 
 import (
 	"context"
@@ -17,8 +17,8 @@ import (
 	"go.opentelemetry.io/collector/extension"
 	"go.uber.org/zap"
 
-	extensionDef "github.com/DataDog/datadog-agent/comp/otelcol/ddflareextension/def"
-	"github.com/DataDog/datadog-agent/comp/otelcol/ddflareextension/impl/internal/metadata"
+	extensionDef "github.com/DataDog/datadog-agent/comp/otelcol/extension/def"
+	"github.com/DataDog/datadog-agent/comp/otelcol/extension/impl/internal/metadata"
 )
 
 // Type exports the internal metadata type for easy reference
