+++ comp/otelcol/extension/impl/testdata/config.yaml
